requirements.txt
python-telegram-bot==20.5
requests>=2.31.0
aiohttp>=3.9.5